Reorder functions and files related to article, abstracts, reviews, sciences and methodologies
